src/bookcover_wallpaper/layout.py
"""Layout algorithms for arranging book covers."""

from pathlib import Path
from PIL import Image


class MasonryLayout:
    """Pinterest-style masonry layout for book covers."""

    def __init__(
        self,
        width: int,
        height: int,
        gap: int = 4,
        aspect_ratio: tuple[int, int] = (2, 3),
    ):
        """Initialize masonry layout.

        Args:
            width: Canvas width in pixels
            height: Canvas height in pixels
            gap: Gap between covers in pixels (default: 4)
            aspect_ratio: Standard book cover ratio (default: 2:3)
        """
        self.width = width
        self.height = height
        self.gap = gap
        self.aspect_ratio = aspect_ratio

    def calculate_layout(self, cover_paths: list[Path]) -> list[tuple[Path, tuple[int, int], tuple[int, int]]]:
        """Calculate positions and sizes for all covers.

        Args:
            cover_paths: List of paths to cover images

        Returns:
            List of (path, (x, y), (w, h)) tuples
        """
        # TODO: Implement masonry algorithm
        # 1. Determine cover width based on canvas width and desired columns
        # 2. Calculate cover height from width using 2:3 aspect ratio
        # 3. Initialize column heights tracker
        # 4. For each cover:
        #    - Place in shortest column
        #    - Update column height
        # 5. Return list of positions and sizes
        return []

    def _calculate_columns(self) -> int:
        """Calculate optimal number of columns for the canvas width."""
        # TODO: Calculate based on width
        # Typical cover width might be 200-300px
        return 6
