pyproject.toml
[project]
name = "bookcover-wallpaper"
version = "0.1.0"
description = "Generate landscape wallpapers from tiled book covers using Goodreads, web search, or local images"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "click>=8.3.1",
    "httpx>=0.28.1",
    "lxml>=6.0.2",
    "pillow>=12.1.0",
    "pydantic>=2.12.5",
]

[project.scripts]
bookcover-wallpaper = "bookcover_wallpaper.cli:main"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true
